All the basic CRUD stuff done

+++ Juris/Api/Services/AppointmentRequestService.cs
@@ -1,7 +1,6 @@
 ﻿using System.Net;
 using Juris.Api.Exceptions;
 using Juris.Data.Repositories;
-using Juris.Models.Constants;
 using Juris.Models.Entities;
 using Juris.Models.Identity;
 using Microsoft.AspNetCore.Identity;
@@ -10,8 +9,8 @@ namespace Juris.Api.Services;
 
 public class AppointmentRequestService : IAppointmentRequestService
 {
-    private readonly IUnitOfWork _unitOfWork;
     private readonly IGenericRepository<AppointmentRequest> _repository;
+    private readonly IUnitOfWork _unitOfWork;
     private readonly UserManager<User> _userManager;
 
     public AppointmentRequestService(IUnitOfWork unitOfWork, UserManager<User> userManager)
@@ -20,13 +19,13 @@ public class AppointmentRequestService : IAppointmentRequestService
         _repository = unitOfWork.AppointmentRequestRepository;
         _userManager = userManager;
     }
-    
-    public async Task<IList<AppointmentRequest>> GetAllRequests(long userId)
+
+    public async Task<IEnumerable<AppointmentRequest>> GetAllRequests(long userId)
     {
         var user = await _userManager.FindByIdAsync(userId.ToString());
         if (user == null)
             throw new HttpResponseException(HttpStatusCode.NotFound, $"User with id={userId} not found");
-        
+
         return await _repository.GetAll(a => a.UserId == userId);
     }
 
@@ -41,21 +40,29 @@ public class AppointmentRequestService : IAppointmentRequestService
         await _unitOfWork.Save();
     }
 
-    public async Task DeleteRequest(long requestId)
+    public async Task DeleteRequest(long requestId, long userId)
     {
         var request = await _repository.GetById(requestId);
         if (request == null)
             throw new HttpResponseException(HttpStatusCode.NotFound, $"Request with id={requestId} not found");
-        
+
+        if (request.UserId != userId)
+            throw new HttpResponseException(HttpStatusCode.Unauthorized,
+                $"Unauthorized to delete request id={requestId}");
+
         await _repository.Delete(requestId);
         await _unitOfWork.Save();
     }
 
-    public async Task UpdateRequestStatus(string status, long requestId)
+    public async Task UpdateRequestStatus(string status, long requestId, long userId)
     {
         var request = await _repository.GetById(requestId);
         if (request == null)
             throw new HttpResponseException(HttpStatusCode.NotFound, $"Request with id={requestId} not found");
+
+        if (request.UserId != userId)
+            throw new HttpResponseException(HttpStatusCode.Unauthorized,
+                $"Unauthorized to modify request id={requestId}");
 
         request.Status = status;
         _repository.Update(request);
